+++ src/main/java/mod/emt/kami/items/tools/ItemIchoriumHoe.java
@@ -1,0 +1,61 @@
+package mod.emt.kami.items.tools;
+
+import mod.emt.kami.Kami;
+import mod.emt.kami.registry.ModItemsKAMI;
+import mod.emt.kami.utils.helpers.ItemHelper;
+import net.minecraft.creativetab.CreativeTabs;
+import net.minecraft.enchantment.Enchantment;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.init.Enchantments;
+import net.minecraft.item.EnumRarity;
+import net.minecraft.item.ItemHoe;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.NonNullList;
+import net.minecraftforge.common.IRarity;
+import org.jetbrains.annotations.NotNull;
+import thaumcraft.api.items.IWarpingGear;
+
+import java.util.Objects;
+
+public class ItemIchoriumHoe extends ItemHoe implements IWarpingGear {
+    public ItemIchoriumHoe() {
+        this("ichorium_hoe");
+    }
+
+    protected ItemIchoriumHoe(String unlocName) {
+        super(ModItemsKAMI.MATERIAL_ICHORIUM);
+        this.setRegistryName(Kami.MOD_ID, unlocName);
+        this.setTranslationKey(Objects.requireNonNull(this.getRegistryName()).toString());
+        this.setCreativeTab(Kami.tabKAMI);
+    }
+
+    @Override
+    public void getSubItems(@NotNull CreativeTabs tab, @NotNull NonNullList<ItemStack> items) {
+        if(this.isInCreativeTab(tab)) {
+            ItemStack stack = new ItemStack(this);
+            ItemHelper.setUnbreakable(stack);
+            items.add(stack);
+        }
+    }
+
+    @Override
+    public boolean isEnchantable(@NotNull ItemStack stack) {
+        return true;
+    }
+
+    @Override
+    public @NotNull IRarity getForgeRarity(@NotNull ItemStack stack) {
+        return EnumRarity.EPIC;
+    }
+
+    @Override
+    public int getWarp(ItemStack itemStack, EntityPlayer entityPlayer) {
+        return 1;
+    }
+
+    @Override
+    public boolean canApplyAtEnchantingTable(@NotNull ItemStack stack, @NotNull Enchantment enchantment) {
+        if (enchantment == Enchantments.MENDING || enchantment == Enchantments.UNBREAKING) return false;
+        return super.canApplyAtEnchantingTable(stack, enchantment);
+    }
+}