+++ src/main/java/mod/emt/kami/registry/ModItemsKAMI.java
@@ -21,6 +21,7 @@ public class ModItemsKAMI {
     public static Item.ToolMaterial MATERIAL_ICHORIUM = EnumHelper.addToolMaterial("ICHORIUM", 5, -1, 10, 4f, 25);
 
     public static Item AWAKENED_ICHORIUM_AXE;
+    public static Item AWAKENED_ICHORIUM_SICKLE;
     public static Item AWAKENED_ICHORIUM_PICKAXE;
     public static Item AWAKENED_ICHORIUM_SHOVEL;
     public static Item AWAKENED_ICHORIUM_SWORD;
@@ -33,6 +34,7 @@ public class ModItemsKAMI {
     public static Item ICHOR;
     public static Item ICHORIUM_AXE;
     public static ItemIchoriumCaster ICHORIUM_CASTER;
+    public static Item ICHORIUM_HOE;
     public static Item ICHORIUM_INGOT;
     public static Item ICHORIUM_NUGGET;
     public static Item ICHORIUM_PICKAXE;
@@ -65,10 +67,12 @@ public class ModItemsKAMI {
         MOD_ITEMS.add(ICHORIUM_SHOVEL = new ItemIchoriumShovel());
         MOD_ITEMS.add(ICHORIUM_PICKAXE = new ItemIchoriumPickaxe());
         MOD_ITEMS.add(ICHORIUM_AXE = new ItemIchoriumAxe());
+        MOD_ITEMS.add(ICHORIUM_HOE = new ItemIchoriumHoe());
         MOD_ITEMS.add(AWAKENED_ICHORIUM_SWORD = new ItemAwakenedSword());
         MOD_ITEMS.add(AWAKENED_ICHORIUM_SHOVEL = new ItemAwakenedShovel());
         MOD_ITEMS.add(AWAKENED_ICHORIUM_PICKAXE = new ItemAwakenedPickaxe());
         MOD_ITEMS.add(AWAKENED_ICHORIUM_AXE = new ItemAwakenedAxe());
+        MOD_ITEMS.add(AWAKENED_ICHORIUM_SICKLE = new ItemAwakenedSickle());
         MOD_ITEMS.add(ICHORIUM_CASTER = new ItemIchoriumCaster());
     }
 }